requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
httpx
google-generativeai
PyPDF2
python-dotenv